docker/classifier/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
pillow==11.0.0
imagehash==4.3.1
transformers==4.46.3
numpy==1.26.4
